changed all logic

+++ pkg/models/subscription.go
@@ -3,5 +3,5 @@ package models
 type Subscription struct {
 	User   User
 	Course Course
-	Price  int
+	Price  float64
 }
